Completely separate the colours
Some feedback we've received has been to add some kind of
indication as to where the opposite actor is at all times,
so that the player can think better. Motivated by that,
I'm instead removing the targets which crossed between
the sides before. So now the orange area is completely
orange, and the blue completely blue.

We'll see how that playtests!

One problem I noticed right away was that the levels where
one actor _must_ push the other onto their target, the solutions
feel weirder. Like, "I just pushed this actor onto an unadorned
space... and then won the level".