[WIP] Changing the Legend
The goal here was to have some levels start blue and others
start orange, but I realized that in order to do this we need
an entirely new set of 8 legend symbols for the blue starting
since 8 is only enough to represent one starting side (four for
wall, four for floor).

After poking once or twice at doing this, I decided that I would
instead start by fixing the Legend so that it is easier to reason
about.

Of course this meant changing the letters A and B everywhere in the
text.